# PERF: Optimize Extensions.cs